+++ Makefile
@@ -51,7 +51,7 @@ purge:
 uninstall:
 	$(call UNLINK_LAB_EXTENSION,@elyra/application)
 	$(call UNLINK_LAB_EXTENSION,@elyra/ui-components)
-	$(call UNLINK_LAB_EXTENSION,@elyra/metadata)
+	$(call UNLINK_LAB_EXTENSION,@elyra/metadata-common)
 	$(call UNINSTALL_LAB_EXTENSION,@elyra/theme-extension)
 	$(call UNINSTALL_LAB_EXTENSION,@elyra/code-snippet-extension)
 	$(call UNINSTALL_LAB_EXTENSION,@elyra/metadata-extension)
@@ -93,10 +93,10 @@ install-server: build-server ## Install backend
 install-ui: build-ui
 	$(call LINK_LAB_EXTENSION,application)
 	$(call LINK_LAB_EXTENSION,ui-components)
-	$(call LINK_LAB_EXTENSION,metadata)
+	$(call LINK_LAB_EXTENSION,metadata-common)
 	$(call INSTALL_LAB_EXTENSION,theme)
 	$(call INSTALL_LAB_EXTENSION,code-snippet)
-	$(call INSTALL_LAB_EXTENSION,metadata-extension)
+	$(call INSTALL_LAB_EXTENSION,metadata)
 	$(call INSTALL_LAB_EXTENSION,pipeline-editor)
 	$(call INSTALL_LAB_EXTENSION,python-editor)
 
@@ -133,7 +133,7 @@ dist-ui: build-ui
 	mkdir -p dist
 	$(call PACKAGE_LAB_EXTENSION,theme)
 	$(call PACKAGE_LAB_EXTENSION,code-snippet)
-	$(call PACKAGE_LAB_EXTENSION,metadata-extension)
+	$(call PACKAGE_LAB_EXTENSION,metadata)
 	$(call PACKAGE_LAB_EXTENSION,pipeline-editor)
 	$(call PACKAGE_LAB_EXTENSION,python-editor)
 	cd dist && curl -o jupyterlab-git-$(GIT_VERSION).tgz $$(npm view @jupyterlab/git@$(GIT_VERSION) dist.tarball) && cd -
