+++ packages/metadata-common/package.json
@@ -1,11 +1,10 @@
 {
-  "name": "@elyra/metadata-extension",
+  "name": "@elyra/metadata-common",
   "version": "1.0.0-rc.1",
-  "description": "JupyterLab extension - View and edit metadata",
+  "description": "JupyterLab - Widgets for interacting with metadata",
   "keywords": [
     "jupyter",
-    "jupyterlab",
-    "jupyterlab-extension"
+    "jupyterlab"
   ],
   "homepage": "https://github.com/elyra-ai/elyra",
   "bugs": {
@@ -31,20 +30,21 @@
     "watch": "tsc -w"
   },
   "dependencies": {
-    "@elyra/metadata": "^1.0.0-rc.1",
+    "@blueprintjs/core": "^3.22.2",
+    "@elyra/application": "^1.0.0-rc.1",
+    "@elyra/ui-components": "^1.0.0-rc.1",
     "@jupyterlab/application": "^2.0.2",
     "@jupyterlab/apputils": "^2.0.2",
     "@jupyterlab/codeeditor": "^2.0.2",
     "@jupyterlab/ui-components": "^2.0.2",
     "@lumino/algorithm": "^1.2.3",
-    "@lumino/widgets": "^1.11.1"
+    "@lumino/messaging": "^1.3.3",
+    "@lumino/widgets": "^1.11.1",
+    "react": "~16.9.0"
   },
   "devDependencies": {
     "rimraf": "~3.0.0",
     "typescript": "~3.7.3"
-  },
-  "jupyterlab": {
-    "extension": true
   },
   "publishConfig": {
     "access": "public"
