+++ packages/metadata-common/src/index.ts
@@ -13,3 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
+import '../style/index.css';
+
+export * from './MetadataEditor';
+export * from './MetadataWidget';
