+++ packages/metadata-common/tsconfig.json
@@ -10,9 +10,6 @@
       "path": "../application"
     },
     {
-      "path": "../metadata"
-    },
-    {
       "path": "../ui-components"
     }
   ]
