+++ packages/metadata/package.json
@@ -1,10 +1,11 @@
 {
-  "name": "@elyra/metadata",
+  "name": "@elyra/metadata-extension",
   "version": "1.0.0-rc.1",
-  "description": "JupyterLab - Widgets for interacting with metadata",
+  "description": "JupyterLab extension - View and edit metadata",
   "keywords": [
     "jupyter",
-    "jupyterlab"
+    "jupyterlab",
+    "jupyterlab-extension"
   ],
   "homepage": "https://github.com/elyra-ai/elyra",
   "bugs": {
@@ -30,21 +31,20 @@
     "watch": "tsc -w"
   },
   "dependencies": {
-    "@blueprintjs/core": "^3.22.2",
-    "@elyra/application": "^1.0.0-rc.1",
-    "@elyra/ui-components": "^1.0.0-rc.1",
+    "@elyra/metadata-common": "^1.0.0-rc.1",
     "@jupyterlab/application": "^2.0.2",
     "@jupyterlab/apputils": "^2.0.2",
     "@jupyterlab/codeeditor": "^2.0.2",
     "@jupyterlab/ui-components": "^2.0.2",
     "@lumino/algorithm": "^1.2.3",
-    "@lumino/messaging": "^1.3.3",
-    "@lumino/widgets": "^1.11.1",
-    "react": "~16.9.0"
+    "@lumino/widgets": "^1.11.1"
   },
   "devDependencies": {
     "rimraf": "~3.0.0",
     "typescript": "~3.7.3"
+  },
+  "jupyterlab": {
+    "extension": true
   },
   "publishConfig": {
     "access": "public"
