+++ packages/metadata/tsconfig.json
@@ -10,6 +10,9 @@
       "path": "../application"
     },
     {
+      "path": "../metadata-common"
+    },
+    {
       "path": "../ui-components"
     }
   ]
